+++ server.js
@@ -5,28 +5,7 @@ var router = express.Router();
 
 
 app.use(express.static('public'));
-
-//* MONGO DB Connection with mongoose * //
-var mongoose = require('mongoose');
-mongoose.connect('mongodb://localhost/mapifyTest');
-
-var db = mongoose.connection;
-db.on('error', console.error.bind(console, 'connection error:'));
-db.once('open', function (callback) {
-  //DO ANYTHING HERE//
-});
-
-//*TWEET SCHEMA AND MODEL FOR MONGO *//
-var tweetSchema = mongoose.Schema({
-  createdAt:    Date,
-  tweetID:      Number,
-  coordinates:  Array,
-  text:         String
-});
-
-//*TO USE TWEETSCHEMA - WE NEED TO CONVERT IT INTO A MODEL*//
-var Tweet = mongoose.model('Tweet', tweetSchema);
-
+var TweetModel = require('./public/js/models/tweet.js');
 
 /*Get home page */
 router.get('/', function (req, res) {
@@ -86,7 +65,7 @@ var T = new Twit({
 // }, 1000);
 
 //* Show saved items in MongoDB *//
-Tweet.find(function(err,tweets){
+TweetModel.find(function(err,tweets){
   if(err) return console.error(err);
   console.log(tweets);
 })
